Do not use token for public resources
When accessing public resources, we do not need to use the token. This
especially has advantages if the token is not usable any longer, e.g.
because it expired or was revoked. In those cases, the launcher would
previously show an even when trying to load the official source, because
it was using the invalid token.