Small enhancements to the 5.2.0 wording adjustments (#1108)
* Correct apostrophe

Because in other places we also do not use &lsquo; but rather &apos;.

* Remove wrong space and commented out code

* Replace own description with reference to docs

The documentation is not deployed yet, but we will publish there what
can already be seen at
https://github.com/NordicSemiconductor/pc-nrfconnect-launcher/blob/main/doc/docs/working_with_authentications_tokens.md#generating-a-new-token

* Correct description

- non-external -> restricted
- updating restricted sources will not lead to errors but just not be done any longer

* Update snapshot

* Provide additional link to Artifactory

In the dialogs provide an additional link directly to the Artifactory
page where folks can create a token. Don't add this to the settings
page, because I want to keep that as concise as possible.

* Add commas back