Fix: Migrating sources
This error was introduced when switching to also supporting short
Artifactory URLs. In the files on the server we still use URLs with the
download API (so they also work with old versions of the launcher). But
when it comes to detect which apps are installed locally, we also
compare these URLs, so they have to be identical in the case we migrate
here.

STR:
1. Run launcher 5.1.0
2. Add Direction Viewer source https://developer.nordicsemi.com/.pc-tools/nrfconnect-apps/directionfinding/source.json
3. Install the Direction Viewer app
4. Run launcher 5.2.0
5. Set token

ER:
- Direction Viewer app shown as still normally installed (and being able
  to reinstall)

AR:
- Direction Viewer app shown as “not available for download anymore”
  (and if it is uninstalled, it vanishes, so it is impossible to install
  it again)