Always copy bundled nrfutil to user data dir
Rationale: If a future launcher contains a nrfutil executable which does
not work properly (like we see coming up for Ubuntu 20.04), we want to
enable users to get back to a working launcher by installing an old
launcher which still contains a working nrfutil executable. This is only
possible if we always copy the bundled nrfutil, even if it is already
there.