Fix apps wrongly shown as withdrawn
When jumping back and forth between launcher versions that use
developer.nordicsemi.com and files.nordicsemi.com, apps may be
incorrectly shown as withdrawn.

This is caused by them coming from different URLs on both servers, e.g.
https://developer.nordicsemi.com/.pc-tools/nrfconnect-apps/release-test/pc-nrfconnect-npm.json
on the old and
https://files.nordicsemi.com/artifactory/swtools/internal/ncd/apps/release-test/pc-nrfconnect-npm.json
on the new one.

The fix here is to only compare the last segment of the URL.