Fix: Missing progress bar on app installs
When using the download API of Artifactory (e.g. with the URL
https://files.nordicsemi.com/ui/api/v1/download?isNativeBrowsing=true&repoKey=swtools&path=external/ncd/apps/official/pc-nrfconnect-npm-1.3.1.tgz)
the server does not return the header content-length, making it
impossible to determine how much of the whole file has been downloaded.
For this longer URLs Artifactory also does not support HEAD requests.

So, to determine the total download size, we first do a HEAD request to
the shorter URL (e.g.
https://files.nordicsemi.com/artifactory/swtools/external/ncd/apps/official/pc-nrfconnect-npm-1.3.1.tgz).

For the HEAD request we use the newer net.fetch API because it
simplifies the implementation.

We should also replace the previous use of the net.request API, but this
is outside of the scope of this change.