Support update channels in auto-update
With this, users can select a different update channel than the default
(`latest`). Intentionally there is no UI for setting the update channel,
this needs to be done manually in the config file.

Currently we have only a single alternative update channel, `rc` for
release candidates. If you want to use it, edit your `config.json` and add the entry `"updateChannel": "rc"`.

If you want to go back to the default channel, remove the `updateChannel`
entry from your `config.json` again.

If you have configured an update channel, it will be displayed in the
launcher UI in the top right corner, so you can easily see which channel
you are currently using.

Your `config.json` can be found in these places:
- macOS: `~/Library/Application Support/nrfconnect/config.json`
- Linux: `~/.config/nrfconnect/config.json` (unless $XDG_CONFIG_HOME is
  set)
- Windows: `%APPDATA%\nrfconnect\config.json`