Interrupt release workflows for Windows signing
We currently must sign the Windows binaries manually before we can
release them.

So, this changes both release workflows (RC and production) and
introduces new workflows for continuing releases. With those, the
release processes will be like this:

1. Run the release workflow (happens automatically for RCs, manually for
   production), which only prepare a draft release on GitHub and then
   stops.

2. Now we candownload the windows binary, sign it, and upload it back to
   the draft release.

3. Run the continue release workflow (There are separate ones for RC and
   production), which will publish the draft release and upload all
   files to Artifactory. With this, the release is completed.

This is only a temporary solution. When we do not need to manually sign
the windows binaries anymore, we will revert this change.