Fix: Error when running binary
`fs.cpSync` failed, when the source was in the asar archive. This only
happened when running the binary, not when running from source because
only in the binary an asar archive is used.

So now the files in nrfutil-sandboxes are unpacked from the asar archive
to `app.asar.unpacked` and then copied from there.