Make launcher more robust against nrfutil failures
Before, when during the app compatibility warning check the invocation
of nrfutil failed, an error was printed to the console but otherwise
nothing happened.

This is especially bad if this happens during launching an app, because
it means that the app will not be launched and the user also does not
get any feedback about it.

Now in this case the check still a warning will be printed to the
console but the opearation will otherwise continue.

This has not occured in production, as far as we know, but it happened
during developing a new nrfutil feature. And there is some chance, that
this might also happen in production, so it is better to make this more
robust.