Fix: App list not shown on the very first launch
To reproduce: Remove [the `nrfconnect/`
folder](https://nordicsemiconductor.github.io/pc-nrfconnect-docs/configuration#file-system-location)
and then launch. When the "Help us improve nRF Connect for Desktop"
dialog is shown, it doesn't matter how it is closed, afterwards the list
of apps was not shown. Only on the next launch it will be shown.

The fix was introduced in PR #1173.